[ci] Improved upstream manifest handling
This change introduces a new `filter` and `exclude` key to the upstream.manifest
section of the settings.yml file. This allows for more fine-grained control over
which files are copied from the upstream repository to the git repository.

This replaces the files and remove arrays in the get-latest-upstream-manifest
script so it doesn't have to be modified on a per-repo basis.